refactor: :rewind: Changed code based on reviews
