+++ package.json
@@ -31,10 +31,5 @@
     "esbuild",
     "sharp"
   ],
-  "dependencies": {
-    "dexie": "^4.0.9",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.2.5",
-    "uuid": "^11.0.3"
-  }
+  "dependencies": {}
 }
